Move the health check tests into util_test.go
A new file under octopusdeploy/ fails the SDK additions check, and these
tests need package access to expandMachineHealthCheckPolicy and
getMachineHealthCheckPolicySchema so they cannot move to the framework
directory.
